+++ index.js
@@ -1,1 +1,3 @@
 export { default as default } from './modules/ReactTimeAgo'
+// "Smart" autoupdate intervals should be moved to `javascript-time-ago`'s grading scale.
+// export { UPDATE_INTERVALS } from './modules/updateIntervals'